fixed issue with fs sync reading
shudder
