+++ views/manual/index.hbs
@@ -1,7 +1,7 @@
 <div class="container">
 
 	<div class="manual-heading">
-		{{{md_partial partial}}}
+		{{{content}}}
 	</div>
 
 </div><!-- /.container -->
